+++ biomodels_qc/tests/test_validation.py
@@ -203,6 +203,17 @@ class ValidationTestCase(unittest.TestCase):
         _, warnings = validation.validate_xml_file(bad_filename)
         self.assertIn('Some of the metadata in', flatten_nested_list_of_strings(warnings))
 
+    def test_validate_scilab_file(self):
+        filename = os.path.join(self.FIXTURE_DIRNAME, 'BIOMD0000000002', 'BIOMD0000000002.sci')
+        errors, warnings = validation.validate_scilab_file(filename)
+        self.assertEqual(errors, [])
+        self.assertEqual(warnings, [])
+
+        filename = os.path.join(self.FIXTURE_DIRNAME, 'BIOMD0000000001', 'BIOMD0000000001.sci')
+        errors, warnings = validation.validate_scilab_file(filename)
+        self.assertIn('is not valid', flatten_nested_list_of_strings(errors))
+        self.assertEqual(warnings, [])
+
     def test_validate_sedml_file(self):
         filename = os.path.join(self.FIXTURE_DIRNAME, 'BIOMD0000000724', 'Theinmozhi_2018.sedml')
         errors, warnings = validation.validate_sedml_file(filename)
@@ -260,6 +271,11 @@ class ValidationTestCase(unittest.TestCase):
         filename = os.path.join(self.FIXTURE_DIRNAME, 'BIOMD0000000001', 'BIOMD0000000001.vcml')
         errors, warnings = validation.validate_vcml_file(filename)
         self.assertEqual(errors, [])
+        self.assertEqual(warnings, [])
+
+        filename = os.path.join(self.FIXTURE_DIRNAME, 'BIOMD0000000006', 'BIOMD0000000006.vcml')
+        errors, warnings = validation.validate_vcml_file(filename)
+        self.assertIn('not represent a valid model', flatten_nested_list_of_strings(errors))
         self.assertEqual(warnings, [])
 
         png_filename = os.path.join(self.FIXTURE_DIRNAME, 'BIOMD0000000692', 'BIOMD0000000692.png')